+++ Simple-Inventory-Managment-System/Product.cs
@@ -9,10 +9,10 @@ namespace Simple_Inventory_Managment_System
     public class Product
     {
         public string name { get; set; }
-        public double price { get; set; }
+        public decimal price { get; set; }
         public int quantity { get; set; }
 
-        public Product(string name, double price, int quantity)
+        public Product(string name, decimal price, int quantity)
         { 
             this.name = name;
             this.price = price;
